# libs/clients/cache-manager-client/openapi.yaml
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8083","description":"Generated server url"}],"paths":{"/api/v1/tokens/":{"post":{"tags":["user-token-controller"],"operationId":"saveRemoteToken","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/SaveToken"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/manager/cache/status":{"post":{"tags":["cache-creator-controller"],"operationId":"changeCacheState","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CacheStatus"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/manager/cache/runtime":{"post":{"tags":["cache-creator-controller"],"operationId":"createCache","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateCacheRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InternalApiRequestCacheMetadata"}}}}}}},"/api/v1/import/jdbc/cache":{"post":{"tags":["import-controller"],"operationId":"createJdbcCache","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateJdbcCacheRequest"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InternalApiRequestObject"}}}}}}},"/api/v1/import/default/cache":{"post":{"tags":["import-controller"],"operationId":"createJdbcCache_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDefaultJdbcCache"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InternalApiRequestObject"}}}}}}},"/api/v1/default/tables/":{"post":{"tags":["jdbc-gateway-controller"],"operationId":"getJdbcTables","parameters":[{"name":"tableName","in":"query","required":false,"schema":{"type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/PostgresJdbcProps"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InternalApiRequestObject"}}}}}}},"/api/v1/{project}":{"get":{"tags":["info-service-rest-controller"],"operationId":"getProjectCache","parameters":[{"name":"project","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InternalApiRequestObject"}}}}}}},"/api/v1/{project}/{cacheName}":{"get":{"tags":["info-service-rest-controller"],"operationId":"getMetadata","parameters":[{"name":"project","in":"path","required":true,"schema":{"type":"string"}},{"name":"cacheName","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InternalApiRequestObject"}}}}}}},"/api/v1/tokens/{token}":{"get":{"tags":["user-token-controller"],"operationId":"saveRemoteToken_1","parameters":[{"name":"token","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}},"delete":{"tags":["user-token-controller"],"operationId":"deleteToken","parameters":[{"name":"token","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"string"}}}}}}},"/api/v1/projects":{"get":{"tags":["info-service-rest-controller"],"operationId":"getProject","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"type":"object","additionalProperties":{"type":"object"}}}}}}}},"/api/v1/manager/cache/{project}/{cacheName}":{"delete":{"tags":["cache-creator-controller"],"operationId":"deleteCache","parameters":[{"name":"project","in":"path","required":true,"schema":{"type":"string"}},{"name":"cacheName","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/InternalApiRequestObject"}}}}}}}},"components":{"schemas":{"SaveToken":{"type":"object","properties":{"token":{"type":"string"},"project":{"type":"string"}}},"CacheMetadataKey":{"type":"object","properties":{"cacheName":{"type":"string"},"project":{"type":"string"},"publicCacheName":{"type":"string"}}},"CacheStatus":{"type":"object","properties":{"key":{"$ref":"#/components/schemas/CacheMetadataKey"},"status":{"type":"string","enum":["READY","CLEANING","BUSY","CREATED"]}}},"CreateCacheRequest":{"type":"object","properties":{"cache":{"$ref":"#/components/schemas/CacheMetadataKey"},"columns":{"type":"object","additionalProperties":{"type":"string"}},"cacheType":{"type":"string","enum":["PERSISTENCE","DATA_TIME_LIMIT"]}}},"CacheMetadata":{"type":"object","properties":{"status":{"type":"string","enum":["READY","CLEANING","BUSY","CREATED"]},"cacheId":{"type":"string"},"cache":{"$ref":"#/components/schemas/CacheMetadataKey"},"type":{"type":"string","enum":["PERSISTENCE","DATA_TIME_LIMIT"]},"columns":{"type":"object","additionalProperties":{"type":"string"}},"message":{"$ref":"#/components/schemas/Message"},"lastUpdateDate":{"type":"string","format":"date-time"},"size":{"type":"integer","format":"int64"}}},"InternalApiRequestCacheMetadata":{"type":"object","properties":{"result":{"$ref":"#/components/schemas/CacheMetadata"},"success":{"type":"boolean"},"message":{"type":"string"}}},"Message":{"type":"object","properties":{"message":{"type":"string"},"date":{"type":"string","format":"date-time"},"lastSuccess":{"type":"boolean"}}},"CreateJdbcCacheRequest":{"type":"object","properties":{"id":{"type":"string"},"projectId":{"type":"string"},"name":{"type":"string"},"cacheName":{"type":"string"},"properties":{"$ref":"#/components/schemas/PostgresJdbcProps"},"query":{"type":"string"}}},"PostgresJdbcProps":{"type":"object","properties":{"url":{"type":"string"},"username":{"type":"string"},"password":{"type":"string"},"driverClassName":{"type":"string"},"schema":{"type":"string"}}},"InternalApiRequestObject":{"type":"object","properties":{"result":{"type":"object"},"success":{"type":"boolean"},"message":{"type":"string"}}},"CreateDefaultJdbcCache":{"type":"object","properties":{"id":{"type":"string"},"projectId":{"type":"string"},"name":{"type":"string"},"tableName":{"type":"string"},"properties":{"$ref":"#/components/schemas/PostgresJdbcProps"}}}}}}